Pull code out of quiz view-model to make a 'question provider'.  This will make the code easier to read.  Update version to 0.8
