+++ JBQQuizMeApp/JBQQuizMeApp.csproj
@@ -19,10 +19,10 @@
 		<ApplicationIdGuid>AF1980A5-F16C-4734-A101-78BC36B26AD1</ApplicationIdGuid>
 
 		<!-- Versions -->
-		<AssemblyVersion>0.7.0.0</AssemblyVersion>
-		<FileVersion>0.7.0.0</FileVersion>		
-		<ApplicationDisplayVersion>0.7.0</ApplicationDisplayVersion>
-		<ApplicationVersion>4</ApplicationVersion>
+		<AssemblyVersion>0.8.0.0</AssemblyVersion>
+		<FileVersion>0.8.0.0</FileVersion>		
+		<ApplicationDisplayVersion>0.8.0</ApplicationDisplayVersion>
+		<ApplicationVersion>5</ApplicationVersion>
 
 		<SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'ios'">14.2</SupportedOSPlatformVersion>
 		<SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'maccatalyst'">14.0</SupportedOSPlatformVersion>
