+++ JBQQuizMeApp/Platforms/Windows/Package.appxmanifest
@@ -5,7 +5,7 @@
   xmlns:rescap="http://schemas.microsoft.com/appx/manifest/foundation/windows10/restrictedcapabilities"
   IgnorableNamespaces="uap rescap">
 
-  <Identity Name="jbq-quiz-me" Publisher="CN=User Name" Version="0.7.0.0" />
+  <Identity Name="jbq-quiz-me" Publisher="CN=User Name" Version="0.8.0.0" />
 
   <Properties>
     <DisplayName>JBQ Quiz Me!</DisplayName>
