+++ JBQQuizMeApp/QuizPage.xaml
@@ -139,6 +139,7 @@
                 Grid.Row="5" Grid.Column="0" Grid.ColumnSpan="4"
                 MinimumHeightRequest="10"
                 Margin="2"
+                HorizontalTextAlignment="Center"
                 Text="{Binding Path=CurrentQuestion.Passage, StringFormat='Passage: {0}', TargetNullValue=' '}"
                 FontSize="Small"
                 HorizontalOptions="Center"
@@ -148,6 +149,7 @@
                 Grid.Row="6" Grid.Column="0" Grid.ColumnSpan="4"
                 MinimumHeightRequest="10"
                 Margin="2"
+                HorizontalTextAlignment="Center"
                 Text="{Binding Path=CurrentQuestion.Number, StringFormat='Bible Fact-Pak™ Question #{0}'}"
                 FontSize="Micro"
                 HorizontalOptions="Center"
